refactor(routes): move restart HTTP + MCP endpoints into routes/restart
Follows the routes/health/ and routes/shutdown/ template: splits POST
/restart and the MCP restart tool into mod.rs/logic.rs/http.rs/mcp.rs, with
both surfaces building on a shared logic::build() that spawns the detached
restart helper and builds the response, instead of each doing it separately.
