+++ src/routes/mod.rs
@@ -6,4 +6,5 @@
 pub mod health;
 pub mod http;
 pub mod mcp;
+pub mod restart;
 pub mod shutdown;
