+++ src/routes/restart/mcp_tests.rs
@@ -1,0 +1,60 @@
+#![allow(
+    clippy::missing_docs_in_private_items,
+    reason = "test helpers and fixtures do not need doc comments"
+)]
+
+use super::MoadimMcp;
+
+fn make_handler() -> MoadimMcp {
+    MoadimMcp::new(
+        crate::routines::new_store(),
+        crate::paths::routines_dir(),
+        0,
+        std::sync::Arc::new(tokio::sync::Notify::new()),
+    )
+}
+
+/// Point `MOADIM_HOME_OVERRIDE` at a fresh, empty temp home for the duration of a test, removing it
+/// on drop, so the detached restart helper's log file doesn't land in the real home. Tests in this
+/// crate run single-threaded per binary, so the global env mutation is safe.
+struct TempHome;
+
+impl TempHome {
+    fn set() -> Self {
+        let dir =
+            std::env::temp_dir().join(format!("moadim-restartmcptest-{}", uuid::Uuid::new_v4()));
+        std::fs::create_dir_all(&dir).expect("create temp home");
+        // SAFETY: single-threaded test execution.
+        unsafe {
+            std::env::set_var("MOADIM_HOME_OVERRIDE", &dir);
+        }
+        Self
+    }
+}
+
+impl Drop for TempHome {
+    fn drop(&mut self) {
+        // SAFETY: single-threaded test execution.
+        unsafe {
+            std::env::remove_var("MOADIM_HOME_OVERRIDE");
+        }
+    }
+}
+
+#[test]
+fn restart_tool_spawns_helper_and_acknowledges() {
+    // The tool spawns a detached `current_exe --background` helper; under the test harness that exe
+    // is the test binary, which rejects `--background` and exits at once, so no real server starts.
+    // TempHome keeps the helper's log file out of the real home.
+    let _home = TempHome::set();
+    let handler = make_handler();
+    let result = handler.restart().unwrap();
+    assert!(!result.is_error.unwrap_or(false));
+    let text = match &result.content[0] {
+        rmcp::model::ContentBlock::Text(txt) => txt.text.clone(),
+        _ => panic!("expected text content"),
+    };
+    let val: serde_json::Value = serde_json::from_str(&text).unwrap();
+    assert_eq!(val["status"], "restarting");
+    assert!(val["helper_pid"].as_u64().unwrap() > 0);
+}
