+++ src/routes/restart/mod.rs
@@ -1,0 +1,13 @@
+//! Restart: shared logic (`logic.rs`), HTTP handler (`http.rs`), and MCP tool (`mcp.rs`,
+//! declared from `routes::mcp` so it can reach `MoadimMcp`'s private state).
+
+pub(crate) mod logic;
+pub use logic::RestartResponse;
+
+#[path = "http.rs"]
+mod http;
+#[allow(
+    unused_imports,
+    reason = "utoipa's OpenApi derive resolves this hidden __path_restart type via crate::routes::restart::__path_restart, generated by #[utoipa::path] on the re-exported handler below"
+)]
+pub use http::{__path_restart, restart};
